Script Debugger 5.1 Clippings
Expansion Tags

A clipping’s text is pasted literally into your script, except for the
following expansion tags which are interpreted intelligently:

[[template:text]]
This tag displays text as a placeholder. Alternatively, wrap the text
in delimiters #~...~#; behind the scenes, that’s what a placeholder’s
text really looks like in a script.
[[selected-lines:default text]]
This tag displays the complete lines containing the script’s current
selection. If the current selection is just an insertion point in an
empty line, default text is used, and is displayed as a placeholder.
[[selection:default text]]
This tag displays the script’s current selection. If the current
selection is just an insertion point, default text is used, and is
displayed as a placeholder.
[[select:text]]
This tag displays text also selects it.
[[user]]
This tag displays as the user’s full name, as shown in the Accounts
preference pane.
[[account]]
This tag displays as the user’s short name, as shown in the Accounts
preference pane.
[[organization]]
This tag displays the user’s company name, as shown in the Company
field in the user’s Address Book entry. (If that field is blank,
displays “MyCompanyName”.)
[[date:format]]
This tag displays the current date and time, where format is an
unquoted strftime() format string. Alternatively, just say [[date]] and
a standard format (mm/dd/yy hh:mm:ss) will be used.
[[VARNAME]]
This tag displays the value of the VARNAME environment variable. So,
for example, [[SHELL]] would be expanded to something like /bin/bash.
Anything in double brackets that doesn’t match one of the preceding tag
types is taken to be the name of an environment variable. If there is
no matching environment variable name, the tag is left unchanged; for
example, [[howdy]] becomes [[howdy]].